basic functionality done

+++ data/src/main/java/io/felipeandrade/data/api/response/LiveCurrenciesResponse.kt
@@ -6,5 +6,5 @@ data class LiveCurrenciesResponse (
     val terms: String,
     val timestamp: Int,
     val source: String,
-    val quotes: HashMap<String, Float>?
+    val quotes: HashMap<String, Double>?
 )
